+++ apps/licences/src/pages/licences/[licenseNumber].tsx
@@ -16,8 +16,7 @@ import { Chip, Link, Skeleton, Stack, Typography } from "@mui/material";
 import { GetServerSideProps } from "next";
 import { useRouter } from "next/router";
 import { serverSideTranslations } from "next-i18next/serverSideTranslations";
-import { useMemo } from "react";
-import { useCallback } from "react";
+import { useCallback, useMemo } from "react";
 import { useTranslation } from "react-i18next";
 
 import { GetLayoutProps } from "../_app";
@@ -61,44 +60,52 @@ export default function LicenceDetails() {
     () => [
       {
         key: "dataAttribuzionePunteggio",
-        render: (_, item) => (
-          <>
-            <Typography
-              color="text.secondary"
-              fontWeight="regular"
-              textAlign="initial"
-              variant="body1"
-            >
-              {new Date(item.dataAttribuzionePunteggio).toLocaleDateString()}
-            </Typography>
-            {item.codiceVerbale ? (
-              <Link
-                color="text.primary"
-                component="button"
-                onClick={() => onDetailClicked?.(item)}
-                sx={{
-                  textAlign: "left",
-                  textDecoration: "underline",
-                }}
-              >
-                {item.descrizioneEventoPunteggio}
-              </Link>
-            ) : (
+        render: (_, item) => {
+          const description = item.descrizioneEventoPunteggio?.toLowerCase();
+          return (
+            <>
               <Typography
                 color="text.secondary"
-                fontWeight="medium"
+                fontWeight="regular"
                 textAlign="initial"
                 variant="body1"
               >
-                {item.descrizioneEventoPunteggio}
+                {new Date(item.dataAttribuzionePunteggio).toLocaleDateString()}
               </Typography>
-            )}
-          </>
-        ),
+              {item.codiceVerbale ? (
+                <Link
+                  color="text.primary"
+                  component="button"
+                  onClick={() => onDetailClicked?.(item)}
+                  sx={{
+                    textAlign: "left",
+                    textDecoration: "underline",
+                    textTransform: "capitalize",
+                  }}
+                >
+                  {description}
+                </Link>
+              ) : (
+                <Typography
+                  color="text.secondary"
+                  fontWeight="medium"
+                  textAlign="initial"
+                  textTransform="capitalize"
+                  variant="body1"
+                >
+                  {description}
+                </Typography>
+              )}
+            </>
+          );
+        },
         title: t("licenceDetails.history.columns.detail"),
         widthFactor: 0.8,
       },
       {
+        columnStyle: {
+          textAlign: "right",
+        },
         key: "dataAttribuzionePunteggio",
         render: (_, item) => (
           <Stack alignItems="end" width="100%">
@@ -109,7 +116,6 @@ export default function LicenceDetails() {
             />
           </Stack>
         ),
-
         title: t("licenceDetails.history.columns.variation"),
       },
     ],
@@ -138,17 +144,14 @@ export default function LicenceDetails() {
   }
 
   return (
-    <>
-      <SectionTitle icon="driveLicense" label={data.numeroPatente} />
-      <Stack my={3} spacing={2}>
-        <LicenceSectionDetails data={data} />
-        <CardInfo
-          bottomContent={<Table columns={columns} rows={rows} />}
-          icon={<Icon fontSize="medium" name="documentText" />}
-          title={t("licenceDetails.history.title")}
-        />
-      </Stack>
-    </>
+    <Stack spacing={2}>
+      <LicenceSectionDetails data={data} />
+      <CardInfo
+        bottomContent={<Table columns={columns} rows={rows} />}
+        icon={<Icon fontSize="medium" name="documentText" />}
+        title={t("licenceDetails.history.title")}
+      />
+    </Stack>
   );
 }