+++ apps/payments/src/lib/bff/business.ts
@@ -22,3 +22,30 @@ export const retrievePayments = async (
     return error;
   }
 };
+
+export const retrievePaymentReceipt = async (
+  additionalDataJWS: string,
+  token: string,
+  paymentRequestId: string,
+  fiscalCode: string,
+) => {
+  try {
+    return await getExternalApiClient().stampaRicevutaTelematica({
+      headers: {
+        "Agid-JWT-TrackingEvidence": additionalDataJWS,
+        Authorization: `Bearer ${token}`,
+        codiceFiscale: fiscalCode,
+      },
+      params: {
+        idRichiestaPagamento: paymentRequestId,
+      },
+    });
+  } catch (error) {
+    // eslint-disable-next-line no-console
+    console.error(
+      `An Error has occurred while retrieving payment receipt, caused by: `,
+      error,
+    );
+    return error;
+  }
+};