+++ apps/payments/src/pages/payments/[requestId].tsx
@@ -1,22 +1,28 @@
 import AppLayout from "@/components/layouts/AppLayout";
-import IUVListDetails from "@/components/payment-details/IUVListDetails";
+import { IUVListDetails } from "@/components/payment-details/IUVListDetails";
 import { PaymentSectionDetails } from "@/components/payment-details/PaymentSectionDetails";
 import { GenericError } from "@/components/shared/GenericError";
 import { Pagamento } from "@/generated/bff-openapi";
+import { usePaymentReceipt } from "@/hooks/usePaymentReceipt";
 import { usePayments } from "@/hooks/usePayments";
-import { CardInfo, SectionTitle } from "@io-ipatente/ui";
+import { PRINT_RECIPT_STATUSES_LIST } from "@/utils/consts";
+import { CardInfo, Icon, SectionTitle } from "@io-ipatente/ui";
+import { Box, Button } from "@mui/material";
 import Skeleton from "@mui/material/Skeleton";
 import Stack from "@mui/material/Stack";
+import { saveAs } from "file-saver";
 import { GetServerSideProps } from "next";
 import { useRouter } from "next/router";
+import { useTranslation } from "next-i18next";
 import { serverSideTranslations } from "next-i18next/serverSideTranslations";
-import { useCallback } from "react";
+import { useCallback, useMemo } from "react";
 
 import { GetLayoutProps } from "../_app";
 
 export default function PaymentDetails() {
   const router = useRouter();
   const requestId = router.query.requestId;
+  const { t } = useTranslation();
 
   const selectPaymentByRequestId = useCallback(
     (data: Pagamento[]) =>
@@ -26,6 +32,33 @@ export default function PaymentDetails() {
 
   const { data, error, isError, isLoading, isRefetching, refetch } =
     usePayments(selectPaymentByRequestId);
+
+  const { refetch: refetchReceipt } = usePaymentReceipt(requestId as string);
+
+  const enableDownload = useMemo(
+    () =>
+      !!data &&
+      !!data?.statoPratica.codice &&
+      PRINT_RECIPT_STATUSES_LIST.includes(data.statoPratica.codice),
+    [data],
+  );
+
+  const downloadFile = useCallback(async () => {
+    const { data: receipt, isError } = await refetchReceipt();
+    if (!isError && receipt?.risultato) {
+      const { ext, file, fileName } = receipt.risultato;
+
+      const byteCharacters = atob(file);
+      const byteNumbers = new Array(byteCharacters.length);
+
+      new Array(byteCharacters.length).fill(null).forEach((_, i) => {
+        byteNumbers[i] = byteCharacters.charCodeAt(i);
+      });
+      const byteArray = new Uint8Array(byteNumbers);
+      const blob = new Blob([byteArray], { type: `application/${ext}` });
+      saveAs(blob, `${fileName}.${ext}`);
+    }
+  }, [refetchReceipt]);
 
   if (isLoading || isRefetching) {
     return (
@@ -50,7 +83,23 @@ export default function PaymentDetails() {
 
   return (
     <>
-      <Stack my={3} spacing={2}>
+      <Stack spacing={2}>
+        <SectionTitle
+          label={data.descrizioneTipoPratica || data.idCarrello.toString()}
+        />
+        {enableDownload && (
+          <Box width="max-content">
+            <Button
+              endIcon={<Icon name="download" />}
+              onClick={downloadFile}
+              size="small"
+              variant="contained"
+            >
+              {t("paymentDetails.receipt.save")}
+            </Button>
+          </Box>
+        )}
+
         <PaymentSectionDetails data={data} />
         <IUVListDetails iuvList={data.listaIuv} />
       </Stack>