+++ packages/core/lib/wrappers/with-jwt-auth-handler.ts
@@ -1,18 +1,20 @@
 import { NextResponse } from "next/server";
-import { NextAuthResult, User } from "next-auth";
+import { User } from "next-auth";
 
+import { AuthRouteHandler } from "../types";
 import { handleUnauthorizedErrorResponse } from "../utils/errors";
-
-type AuthParams = Parameters<NextAuthResult["auth"]>;
 
 export const withJWTAuthHandler =
   (
     handler: (
       request: Request,
-      context: { user: User },
+      context: {
+        params?: Record<string, string | string[]>;
+        user: User;
+      },
     ) => Promise<NextResponse> | Promise<Response>,
-  ): AuthParams[0] =>
-  (request) => {
+  ): AuthRouteHandler =>
+  (request, ctx) => {
     const { auth } = request;
 
     if (!auth) {
@@ -20,6 +22,7 @@ export const withJWTAuthHandler =
     }
 
     return handler(request, {
+      ...ctx,
       user: auth.user,
     });
   };