+++ packages/core/lib/wrappers/with-jwt-auth-voucher-handler.ts
@@ -1,22 +1,22 @@
 import { NextResponse } from "next/server";
-import { NextAuthResult, User } from "next-auth";
+import { User } from "next-auth";
 
 import { Voucher } from "../interop/voucher";
+import { AuthRouteHandler } from "../types";
 import { withJWTAuthHandler } from "./with-jwt-auth-handler";
 import { withVoucherHandler } from "./with-voucher-handler";
-
-type AuthParams = Parameters<NextAuthResult["auth"]>;
 
 export const withJWTAuthAndVoucherHandler = (
   handler: (
     request: Request,
     context: {
       additionalDataJWS: string;
+      params?: Record<string, string | string[]>;
       user: User;
       voucher: Voucher;
     },
   ) => Promise<NextResponse> | Promise<Response>,
-): AuthParams[0] =>
+): AuthRouteHandler =>
   withJWTAuthHandler((request, jwtContext) =>
     withVoucherHandler(
       (request, voucherContext) =>
@@ -25,5 +25,5 @@ export const withJWTAuthAndVoucherHandler = (
           ...voucherContext,
         }),
       jwtContext.user.fiscalCode,
-    )(request),
+    )(request, jwtContext),
   );