+++ apps/payments/CHANGELOG.md
@@ -1,5 +1,157 @@
 # io-ipatente-payments
 
+## 0.3.16
+
+### Patch Changes
+
+- Updated dependencies [b256e78]
+  - @io-ipatente/core@0.5.15
+
+## 0.3.15
+
+### Patch Changes
+
+- Updated dependencies [538bb64]
+  - @io-ipatente/core@0.5.14
+
+## 0.3.14
+
+### Patch Changes
+
+- Updated dependencies [e532ce2]
+  - @io-ipatente/core@0.5.13
+
+## 0.3.13
+
+### Patch Changes
+
+- Updated dependencies [95bff35]
+  - @io-ipatente/core@0.5.12
+
+## 0.3.12
+
+### Patch Changes
+
+- Updated dependencies [5a5d1b3]
+  - @io-ipatente/core@0.5.11
+
+## 0.3.11
+
+### Patch Changes
+
+- Updated dependencies [7462e0d]
+  - @io-ipatente/core@0.5.10
+
+## 0.3.10
+
+### Patch Changes
+
+- Updated dependencies [80afcfc]
+  - @io-ipatente/core@0.5.9
+
+## 0.3.9
+
+### Patch Changes
+
+- Updated dependencies [f64dafe]
+  - @io-ipatente/core@0.5.8
+
+## 0.3.8
+
+### Patch Changes
+
+- Updated dependencies [855dd29]
+  - @io-ipatente/core@0.5.7
+
+## 0.3.7
+
+### Patch Changes
+
+- Updated dependencies [0906d71]
+  - @io-ipatente/core@0.5.6
+
+## 0.3.6
+
+### Patch Changes
+
+- Updated dependencies [f041e5e]
+- Updated dependencies [58e8532]
+  - @io-ipatente/core@0.5.5
+  - @io-ipatente/ui@0.8.2
+
+## 0.3.5
+
+### Patch Changes
+
+- Updated dependencies [f0db89c]
+  - @io-ipatente/core@0.5.4
+
+## 0.3.4
+
+### Patch Changes
+
+- Updated dependencies [634dc6a]
+  - @io-ipatente/core@0.5.3
+
+## 0.3.3
+
+### Patch Changes
+
+- Updated dependencies [0371c68]
+  - @io-ipatente/core@0.5.2
+
+## 0.3.2
+
+### Patch Changes
+
+- 35c809a: Middleware refinement
+- Updated dependencies [35c809a]
+- Updated dependencies [fa6b274]
+- Updated dependencies [1727b82]
+  - @io-ipatente/core@0.5.1
+  - @io-ipatente/ui@0.8.1
+
+## 0.3.1
+
+### Patch Changes
+
+- 35dd9d5: Mock Prod config
+
+## 0.3.0
+
+### Minor Changes
+
+- 7e84a75: add download file endpoint
+
+### Patch Changes
+
+- 0dc4ec2: Disable MSW to test memory consumption
+- Updated dependencies [7e84a75]
+  - @io-ipatente/core@0.5.0
+  - @io-ipatente/ui@0.8.0
+
+## 0.2.0
+
+### Minor Changes
+
+- 943b6e3: Summary and detail page for payments merged
+
+### Patch Changes
+
+- b542630: handle nullable fields returned by e-services
+- cdce8d0: Refinement of Assistance page
+- ef2b5bf: practices list and detail
+- c7257ed: Move auth callback route to @io-ipatente/core
+- 12298f7: Show badge licence points in the licenses page
+- Updated dependencies [943b6e3]
+- Updated dependencies [cdce8d0]
+- Updated dependencies [ef2b5bf]
+- Updated dependencies [a0ddaea]
+- Updated dependencies [c7257ed]
+- Updated dependencies [12298f7]
+  - @io-ipatente/ui@0.7.0
+  - @io-ipatente/core@0.4.8
+
 ## 0.1.0
 
 ### Minor Changes